PIM-4395: move mass edit processor
