+++ spec/Pim/Bundle/EnrichBundle/Connector/Processor/MassEdit/UpdateProductValueProcessorSpec.php
@@ -1,17 +1,12 @@
 <?php
 
-namespace spec\Pim\Bundle\EnrichBundle\Processor\MassEdit;
+namespace spec\Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Component\StorageUtils\Updater\PropertySetterInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AttributeRepository;
-use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
-use Pim\Bundle\CatalogBundle\Model\FamilyInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
-use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\ProductMassActionRepositoryInterface;
 use Pim\Component\Connector\Model\JobConfigurationInterface;
 use Pim\Component\Connector\Repository\JobConfigurationRepositoryInterface;
 use Prophecy\Argument;
@@ -20,40 +15,23 @@ use Symfony\Component\Validator\ConstraintViolationList;
 use Symfony\Component\Validator\ConstraintViolationListInterface;
 use Symfony\Component\Validator\ValidatorInterface;
 
-class EditCommonAttributesProcessorSpec extends ObjectBehavior
+class UpdateProductValueProcessorSpec extends ObjectBehavior
 {
     function let(
         PropertySetterInterface $propertySetter,
         ValidatorInterface $validator,
-        ProductMassActionRepositoryInterface $massActionRepository,
-        AttributeRepositoryInterface $attributeRepository,
         JobConfigurationRepositoryInterface $jobConfigurationRepo
     ) {
         $this->beConstructedWith(
             $propertySetter,
             $validator,
-            $massActionRepository,
-            $attributeRepository,
             $jobConfigurationRepo
         );
     }
 
-    function it_returns_the_configuration_fields()
-    {
-        $this->getConfigurationFields()->shouldReturn([]);
-    }
-
-    function it_sets_the_step_execution(StepExecution $stepExecution)
-    {
-        $this->setStepExecution($stepExecution)->shouldReturn($this);
-    }
-
-    function it_sets_values_to_attributes(
-        $validator,
+    function it_sets_values_to_product(
         $propertySetter,
-        FamilyInterface $family,
-        AttributeInterface $attribute,
-        AttributeRepository $attributeRepository,
+        $validator,
         ProductInterface $product,
         StepExecution $stepExecution,
         JobConfigurationRepositoryInterface $jobConfigurationRepo,
@@ -72,27 +50,22 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
                 ]
             )
         );
-
         $violations = new ConstraintViolationList([]);
         $validator->validate($product)->willReturn($violations);
-
-        $attributeRepository->findOneBy(['code' => 'categories'])->willReturn($attribute);
-        $family->hasAttribute($attribute)->willReturn(true);
-        $product->getFamily()->willReturn($family);
         $stepExecution->incrementSummaryInfo('mass_edited')->shouldBeCalled();
 
-        $propertySetter->setData($product, 'categories', ['office', 'bedroom'], [])->shouldBeCalled();
+        $propertySetter->setData($product, 'categories', ['office', 'bedroom'])->shouldBeCalled();
+
+        $this->setStepExecution($stepExecution);
 
         $this->process($product);
     }
 
-    function it_sets_invalid_values_to_attributes(
-        $validator,
+    function it_sets_invalid_values_to_product(
         $propertySetter,
-        FamilyInterface $family,
-        AttributeInterface $attribute,
-        AttributeRepository $attributeRepository,
+        $validator,
         ProductInterface $product,
+        StepExecution $stepExecution,
         ConstraintViolationListInterface $violations,
         StepExecution $stepExecution,
         JobConfigurationRepositoryInterface $jobConfigurationRepo,
@@ -110,21 +83,26 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
                 ]
             )
         );
-
         $validator->validate($product)->willReturn($violations);
         $violation = new ConstraintViolation('error2', 'spec', [], '', '', $product);
         $violations = new ConstraintViolationList([$violation, $violation]);
         $validator->validate($product)->willReturn($violations);
 
-        $attributeRepository->findOneBy(['code' => 'categories'])->willReturn($attribute);
-        $family->hasAttribute($attribute)->willReturn(true);
-        $product->getFamily()->willReturn($family);
-
-        $propertySetter->setData($product, 'categories', ['office', 'bedroom'], [])->shouldBeCalled();
+        $propertySetter->setData($product, 'categories', ['office', 'bedroom'])->shouldBeCalled();
         $this->setStepExecution($stepExecution);
         $stepExecution->addWarning(Argument::cetera())->shouldBeCalled();
         $stepExecution->incrementSummaryInfo('skipped_products')->shouldBeCalled();
 
         $this->process($product);
+    }
+
+    function it_returns_the_configuration_fields()
+    {
+        $this->getConfigurationFields()->shouldReturn([]);
+    }
+
+    function it_sets_the_step_execution(StepExecution $stepExecution)
+    {
+        $this->setStepExecution($stepExecution)->shouldReturn($this);
     }
 }
