+++ src/Pim/Bundle/EnrichBundle/Connector/Processor/MassEdit/UpdateProductValueProcessor.php
@@ -1,41 +1,42 @@
 <?php
 
-namespace Pim\Bundle\EnrichBundle\Processor\MassEdit;
+namespace Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit;
 
-use Akeneo\Component\StorageUtils\Updater\PropertyAdderInterface;
+use Akeneo\Component\StorageUtils\Updater\PropertySetterInterface;
 use Pim\Bundle\CatalogBundle\Exception\InvalidArgumentException;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
+use Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\AbstractMassEditProcessor;
 use Pim\Component\Connector\Repository\JobConfigurationRepositoryInterface;
 use Symfony\Component\Validator\ValidatorInterface;
 
 /**
- * Processor to add product value in a mass edit
+ * Processor to update product value in a mass edit
  *
  * @author    Olivier Soulet <olivier.soulet@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AddProductValueProcessor extends AbstractMassEditProcessor
+class UpdateProductValueProcessor extends AbstractMassEditProcessor
 {
-    /** @var PropertyAdderInterface */
-    protected $propertyAdder;
+    /** @var PropertySetterInterface */
+    protected $propertySetter;
 
     /** @var ValidatorInterface */
     protected $validator;
 
     /**
-     * @param PropertyAdderInterface              $propertyAdder
+     * @param PropertySetterInterface             $propertySetter
      * @param ValidatorInterface                  $validator
      * @param JobConfigurationRepositoryInterface $jobConfigurationRepo
      */
     public function __construct(
-        PropertyAdderInterface $propertyAdder,
+        PropertySetterInterface $propertySetter,
         ValidatorInterface $validator,
         JobConfigurationRepositoryInterface $jobConfigurationRepo
     ) {
         parent::__construct($jobConfigurationRepo);
 
-        $this->propertyAdder = $propertyAdder;
+        $this->propertySetter = $propertySetter;
         $this->validator      = $validator;
     }
 
@@ -52,7 +53,7 @@ class AddProductValueProcessor extends AbstractMassEditProcessor
 
         $actions = $configuration['actions'];
 
-        $this->addData($product, $actions);
+        $this->setData($product, $actions);
 
         if (null === $product || (null !== $product && !$this->isProductValid($product))) {
             $this->stepExecution->incrementSummaryInfo('skipped_products');
@@ -81,17 +82,17 @@ class AddProductValueProcessor extends AbstractMassEditProcessor
     }
 
     /**
-     * Add data from $actions to the given $product
+     * Set data from $actions to the given $product
      *
      * @param ProductInterface $product
      * @param array            $actions
      *
-     * @return AddProductValueProcessor
+     * @return UpdateProductValueProcessor
      */
-    protected function addData(ProductInterface $product, array $actions)
+    protected function setData(ProductInterface $product, array $actions)
     {
         foreach ($actions as $action) {
-            $this->propertyAdder->addData($product, $action['field'], $action['value']);
+            $this->propertySetter->setData($product, $action['field'], $action['value']);
         }
 
         return $this;
